"less is more"
