--- test_prompt.py
@@ -1,7 +1,0 @@
-import promptlib
-
-prompter = promptlib.Files()
-
-dir = prompter.dir()
-
-print(dir, " ")
